Simplify FEATURES_IN manipulation.
After successful features configuration upload, the pointers registers
are already set to the FEATURES_IN area, so we don't really need to
touch them, and we can update this configurations just reading the
whole 70 bytes and writing them back again with our changes. This
simplifies the code and it's really not an issue since it only
happens at configuration stage.